+++ apps/web/app/(auth-pages)/login/page.tsx
@@ -1,5 +1,3 @@
-import { redirect } from 'next/navigation';
-
 import {
   LoginTitle,
   EmailInput,
@@ -7,25 +5,14 @@ import {
   LoginButton,
   SignupInfo,
 } from '../../../components/login';
-import { setServerSession } from '../../../services/login/setServerSession';
+import { loginAction } from '../../../server-actions/login/loginAction';
 
 const LoginPage = () => {
-  async function handleLogin(formData: FormData) {
-    'use server';
-
-    const email = formData.get('email') as string;
-    const password = formData.get('password') as string;
-
-    await setServerSession(email, password);
-
-    redirect('/');
-  }
-
   return (
     <main className="flex min-h-screen items-center justify-center">
       <div className="w-full max-w-[375px] px-4">
         <LoginTitle />
-        <form action={handleLogin}>
+        <form action={loginAction}>
           <div className="space-y-4 pt-4 pb-12">
             <EmailInput />
             <PasswordInput />
